+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## Version 0.3 (working)
 * Support for `expires` on `write` calls
+* Add `write_batch` API
 
 ## Version 0.2 (29-Apr-2026)
 * Add `space_id` and `point_types` options to `points()` and `values()`